feature: 함수 py파일 작성
